requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31

nltk>=3.8
vaderSentiment>=3.3
textblob>=0.17

scikit-learn>=1.3

python-dotenv>=1.0

transformers>=4.40
torch>=2.2

matplotlib>=3.8

groq>=0.9

langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.0.3

sentence-transformers>=2.7

faiss-cpu>=1.7
reportlab
google-search-results
fastapi
uvicorn
python-multipart
pytz